+++ presentation/src/main/java/com/bookmark/presentation/features/club/adapter/HomeDiffUtilCallback.kt
@@ -1,4 +1,4 @@
-package com.bookmark.presentation.features.home.adapter
+package com.bookmark.presentation.features.club.adapter
 
 import androidx.recyclerview.widget.DiffUtil
 import com.example.bookmark.domain.model.Book
