+++ presentation/src/main/java/com/bookmark/presentation/features/profile/comment/adapter/CommentDiffUtilCallback.kt
@@ -1,4 +1,4 @@
-package com.bookmark.presentation.features.comment.adapter
+package com.bookmark.presentation.features.profile.comment.adapter
 
 import androidx.recyclerview.widget.DiffUtil
 import com.example.bookmark.domain.model.Comment
